global app state first version added
also project donations fetch working on table, still needs to stylize
