+++ src/components/project/index.js
@@ -1,1 +1,2 @@
 export { DonationsTab } from './donationsTab'
+export { ProjectDonatorView } from './donatorView'
